+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/datagen/ModDataMapProvider.kt
@@ -1,0 +1,20 @@
+package dev.aaronhowser.mods.irregular_implements.datagen
+
+import dev.aaronhowser.mods.irregular_implements.registry.ModItems
+import net.minecraft.core.HolderLookup
+import net.minecraft.data.PackOutput
+import net.neoforged.neoforge.common.data.DataMapProvider
+import net.neoforged.neoforge.registries.datamaps.builtin.Compostable
+import net.neoforged.neoforge.registries.datamaps.builtin.NeoForgeDataMaps
+import java.util.concurrent.CompletableFuture
+
+class ModDataMapProvider(
+	output: PackOutput,
+	lookupProvider: CompletableFuture<HolderLookup.Provider>
+) : DataMapProvider(output, lookupProvider) {
+
+	override fun gather(provider: HolderLookup.Provider) {
+		builder(NeoForgeDataMaps.COMPOSTABLES)
+			.add(ModItems.BEAN.id, Compostable(0.3f), false)
+	}
+}