+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Fixed
 
 - The Advanced Redstone Interface no longer stops working when the chunk loads
+- The Collapse Effect now works on mobs! It makes them walk backwards
 - Plate models don't use the Pressure Plate model as a base any more, so it isn't cutting off the outermost 2 pixels
 
 # 1.10.5